+++ be/src/exec/hdfs-scan-node-base.h
@@ -223,7 +223,7 @@ class HdfsScanNodeBase : public ScanNode {
   /// Old API for compatibility with text scanners (e.g. LZO text scanner).
   io::ScanRange* AllocateScanRange(hdfsFS fs, const char* file, int64_t len,
       int64_t offset, int64_t partition_id, int disk_id, bool try_cache,
-      bool expected_local, int mtime, const io::ScanRange* original_split = NULL);
+      bool expected_local, int64_t mtime, const io::ScanRange* original_split = NULL);
 
   /// Adds ranges to the io mgr queue. 'num_files_queued' indicates how many file's scan
   /// ranges have been added completely.  A file's scan ranges are added completely if no
